+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Dynamic type checking of initialize method
 - Bug with handling constants of certain types
 - Broken rdl_attr_* methods
+- Bug handling optional annotated/dependent types
 - Add RDLAnnotate to rdl_disable
 
 ## [2.1.0] - 2017-06-14